trying to refer to latest relevant specs
XPath v2 and XQuery v1 seem to have a mangled set of references because the w3c site is trying to drive readers to the latest versions. Whether or not actual functionality is backwards compatible, the URL notation does not seem to be. e.g. "##dayTimeDuration" is not a valid target in the newer versions.